Add virtual environment activation and serial communication
This commit adds code to activate the appropriate virtual environment based on the operating system. It also includes code to establish a serial connection and read data from a serial device. Additionally, a `.gitignore` file is created to exclude the virtual environment directory. The required package `pyserial` is added to the `requirements.txt` file.